+++ tests/testthat/test-cctx.R
@@ -4,7 +4,7 @@
 
 test_that("cctx with zstd_serialize works", {
   
-  cctx <- init_zstd_cctx()
+  cctx <- zstd_cctx()
   expect_identical(
     zstd_serialize(mtcars),
     zstd_serialize(mtcars, cctx = cctx)
@@ -22,7 +22,7 @@ test_that("cctx with zstd_serialize works", {
 
 test_that("cctx with zstd_compress works", {
   
-  cctx <- init_zstd_cctx()
+  cctx <- zstd_cctx()
   expect_identical(
     zstd_compress(serialize(mtcars, NULL)),
     zstd_compress(serialize(mtcars, NULL) , cctx = cctx)